BACKUPS/barcodeSystem - backup/requirements.txt
Flask==3.0.0
pandas==2.3.2
openpyxl==3.1.5
reportlab==4.2.2
Pillow==10.4.0
requests==2.32.3
gunicorn==21.2.0